Crate atomicwrites

Re-exports

pub use OverwriteBehavior::AllowOverwrite;

pub use OverwriteBehavior::DisallowOverwrite;

Structs

AtomicFile

Create a file and write to it atomically, in a callback.

Enums

Error

Represents an error raised by AtomicFile.write.

OverwriteBehavior

Whether to allow overwriting if the target file exists.

Functions

move_atomic

Move src to dst. An error will be returned if dst exists.

replace_atomic

Move src to dst. If dst exists, it will be silently overwritten.