Struct atomicwrites::AtomicFile

pub struct AtomicFile { /* fields omitted */ }

Methods

impl AtomicFile

pub fn new<P: AsRef<Path>>(path: P, overwrite: OverwriteBehavior) -> Self

Helper for writing to path in write-only mode.

If DisallowOverwrite is given, errors will be returned from self.write(...) if the file exists.

pub fn new_with_tmpdir<P: AsRef<Path>>(
path: P,
overwrite: OverwriteBehavior,
tmpdir: P
) -> Self

pub fn path(&self) -> &Path

Get the target filepath.

pub fn write<T, E, F>(&self, f: F) -> Result<T, Error<E>> where
F: FnOnce(&mut File) -> Result<T, E>,

Open a temporary file, call f on it (which is supposed to write to it), then move the file atomically to self.path.